+++ lib/src/base_helper.dart
@@ -29,37 +29,39 @@ class BaseHelper {
   ///Also works nice with flutters paginated table.
   Future<TypedResultList<T>> search<T extends Object>(
     String collection, {
-    required SearchParams params,
-    required List<String> searchableColumns,
+    required String searchQuery,
+    required List<String> searchableFields,
     required RecordMapper<T> mapper,
-    List<String>? otherFilters,
-    Map<String, dynamic>? otherParams,
+    int page = 1,
+    int perPage = 30,
+    String? sort,
+    List<String>? additionalExpressions,
+    Map<String, dynamic>? additionalParams,
+    List<String>? fields,
     Map<String, String>? expansions,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) async {
-    String? filter;
-    if (params.query != null || otherFilters != null) {
-      final (
-        String template,
-        Map<String, dynamic> values,
-      ) = HelperUtils.buildQuery(
-        params.query ?? '',
-        searchableColumns,
-        otherFilters: otherFilters,
-        otherParams: otherParams,
-      );
+    final (
+      String template,
+      Map<String, dynamic> values,
+    ) = HelperUtils.buildQuery(
+      searchQuery,
+      searchableFields,
+      otherFilters: additionalExpressions,
+      otherParams: additionalParams,
+    );
 
-      filter = pb.filter(template, values);
-    }
+    final filter = pb.filter(template, values);
 
     final result = await pb
         .collection(collection)
         .getList(
           filter: filter,
-          sort: HelperUtils.getSortOrderFor(params),
-          page: params.page,
-          perPage: params.perPage,
+          sort: sort,
+          page: page,
+          perPage: perPage,
+          fields: fields?.join(','),
           expand: HelperUtils.buildExpansionString(expansions),
           query: query ?? const {},
           headers: headers ?? const {},
@@ -90,6 +92,7 @@ class BaseHelper {
     int perPage = 30,
     bool skipTotal = false,
     String? sort,
+    List<String>? fields,
     Map<String, String>? expansions,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
@@ -108,6 +111,7 @@ class BaseHelper {
           skipTotal: skipTotal,
           expand: HelperUtils.buildExpansionString(expansions),
           sort: sort,
+          fields: fields?.join(','),
           query: query ?? const {},
           headers: headers ?? const {},
         );
@@ -135,6 +139,7 @@ class BaseHelper {
     String? expr,
     Map<String, dynamic>? params,
     String? sort,
+    List<String>? fields,
     Map<String, String>? expansions,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
@@ -146,9 +151,10 @@ class BaseHelper {
     final result = await pb
         .collection(collection)
         .getFullList(
-          fields: filter,
+          filter: filter,
           batch: batch,
           sort: sort,
+          fields: fields?.join(','),
           expand: HelperUtils.buildExpansionString(expansions),
           query: query ?? const {},
           headers: headers ?? const {},
@@ -168,6 +174,7 @@ class BaseHelper {
     String collection, {
     required String id,
     required RecordMapper<T> mapper,
+    List<String>? fields,
     Map<String, String>? expansions,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
@@ -177,6 +184,7 @@ class BaseHelper {
         .getOne(
           id,
           expand: HelperUtils.buildExpansionString(expansions),
+          fields: fields?.join(','),
           query: query ?? const {},
           headers: headers ?? const {},
         );
@@ -191,6 +199,7 @@ class BaseHelper {
     String collection, {
     required String id,
     required RecordMapper<T> mapper,
+    List<String>? fields,
     Map<String, String>? expansions,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
@@ -201,6 +210,7 @@ class BaseHelper {
           page: 1,
           perPage: 1,
           filter: pb.filter('id={:id}', {'id': id}),
+          fields: fields?.join(','),
           expand: HelperUtils.buildExpansionString(expansions),
           query: query ?? const {},
           headers: headers ?? const {},
@@ -223,15 +233,18 @@ class BaseHelper {
     String collection, {
     required Map<String, dynamic> data,
     required RecordMapper<T> mapper,
+
     Map<String, String>? expansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) async {
     final record = await pb
         .collection(collection)
         .create(
-          body: HelperUtils.creationHook(collection, pb, data),
+          body: HelperUtils.preCreationHook(collection, pb, data),
           expand: HelperUtils.buildExpansionString(expansions),
+          fields: fields?.join(','),
           query: query ?? const {},
           headers: headers ?? const {},
         );
@@ -247,6 +260,7 @@ class BaseHelper {
     required T record,
     required RecordMapper<T> mapper,
     Map<String, String>? expansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) async {
@@ -254,8 +268,9 @@ class BaseHelper {
         .collection(collection)
         .update(
           record.id,
-          body: HelperUtils.updateHook(collection, pb, record.toMap()),
+          body: HelperUtils.preUpdateHook(collection, pb, record.toMap()),
           expand: HelperUtils.buildExpansionString(expansions),
+          fields: fields?.join(','),
           query: query ?? const {},
           headers: headers ?? const {},
         );
@@ -295,6 +310,7 @@ class BaseHelper {
     required Map<String, Uint8List> files,
     required RecordMapper<T> mapper,
     Map<String, String>? expansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) async {
@@ -306,6 +322,7 @@ class BaseHelper {
             for (final file in files.entries)
               MultipartFile.fromBytes('files+', file.value, filename: file.key),
           ],
+          fields: fields?.join(','),
           expand: HelperUtils.buildExpansionString(expansions),
           query: query ?? const {},
           headers: headers ?? const {},
@@ -325,6 +342,7 @@ class BaseHelper {
     required List<String> fileNames,
     required RecordMapper<T> mapper,
     Map<String, String>? expansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) async {
@@ -334,6 +352,7 @@ class BaseHelper {
           id,
           body: {'files-': fileNames},
           expand: HelperUtils.buildExpansionString(expansions),
+          fields: fields?.join(','),
           query: query ?? const {},
           headers: headers ?? const {},
         );