+++ lib/src/collection_helper.dart
@@ -20,7 +20,7 @@ class CollectionHelper<T extends PocketBaseRecord> {
     this.pb, {
     required String collection,
     required T Function(Map<String, dynamic>) mapper,
-    this.baseExpansions,
+    this.expansions,
   }) : _mapper = mapper,
        collectionName = collection,
        _helper = BaseHelper(pb);
@@ -62,7 +62,7 @@ class CollectionHelper<T extends PocketBaseRecord> {
   ///   "message": "Example message...",
   ///}
   ///```
-  ///when [baseExpansions] is:
+  ///when [expansions] is:
   ///```
   ///{
   ///   "user_id" : "user"
@@ -74,7 +74,7 @@ class CollectionHelper<T extends PocketBaseRecord> {
   ///final User user;
   ///```
   ///in your model class
-  final Map<String, String>? baseExpansions;
+  final Map<String, String>? expansions;
 
   final BaseHelper _helper;
 
@@ -84,7 +84,7 @@ class CollectionHelper<T extends PocketBaseRecord> {
 
   ///Execute a search on records based on requested table data.
   ///Takes [SearchParams] and fetches a [TypedResultList].
-  ///Internally this method contructs and advanced filter, that keyword searches all the [searchableColumns]
+  ///Internally this method contructs and advanced filter, that keyword searches all the [searchableFields]
   ///for the query provides in [params].
   ///
   ///If commas are present in this query the query will be comma seperated into multiple keywords
@@ -92,22 +92,29 @@ class CollectionHelper<T extends PocketBaseRecord> {
   ///
   ///This method was designed with tables in mind and is especially to make ui tables with searchbars
   Future<TypedResultList<T>> search({
-    required SearchParams params,
-    required List<String> searchableColumns,
-    List<String>? otherFilters,
-    Map<String, dynamic>? otherParams,
+    required String searchQuery,
+    required List<String> searchableFields,
+    int page = 1,
+    int perPage = 30,
+    List<String>? additionalExpressions,
+    Map<String, dynamic>? additionalParams,
+    String? sort,
+    List<String>? fields,
     Map<String, String>? additionalExpansions,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) => _helper.search(
     collectionName,
-    params: params,
-    searchableColumns: searchableColumns,
+    searchQuery: searchQuery,
+    searchableFields: searchableFields,
+    page: page,
+    sort: sort,
+    perPage: perPage,
     mapper: _mapper,
-    otherFilters: otherFilters,
-    otherParams: otherParams,
-    expansions:
-        (baseExpansions ?? {})..addAll(additionalExpansions ?? const {}),
+    additionalExpressions: additionalExpressions,
+    additionalParams: additionalParams,
+    fields: fields,
+    expansions: (expansions ?? {})..addAll(additionalExpansions ?? const {}),
     query: query,
     headers: headers,
   );
@@ -121,6 +128,7 @@ class CollectionHelper<T extends PocketBaseRecord> {
     bool skipTotal = false,
     String? sort,
     Map<String, String>? additionalExpansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) => _helper.getList(
@@ -131,9 +139,9 @@ class CollectionHelper<T extends PocketBaseRecord> {
     page: page,
     perPage: perPage,
     skipTotal: skipTotal,
-    expansions:
-        (baseExpansions ?? {})..addAll(additionalExpansions ?? const {}),
     sort: sort,
+    expansions: (expansions ?? {})..addAll(additionalExpansions ?? const {}),
+    fields: fields,
     query: query,
     headers: headers,
   );
@@ -145,6 +153,7 @@ class CollectionHelper<T extends PocketBaseRecord> {
     Map<String, dynamic>? params,
     String? sort,
     Map<String, String>? additionalExpansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) => _helper.getFullList(
@@ -153,8 +162,8 @@ class CollectionHelper<T extends PocketBaseRecord> {
     expr: expr,
     params: params,
     sort: sort,
-    expansions:
-        (baseExpansions ?? {})..addAll(additionalExpansions ?? const {}),
+    fields: fields,
+    expansions: (expansions ?? {})..addAll(additionalExpansions ?? const {}),
     query: query,
     headers: headers,
   );
@@ -163,14 +172,15 @@ class CollectionHelper<T extends PocketBaseRecord> {
   Future<T> getSingle(
     String id, {
     Map<String, String>? additionalExpansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) => _helper.getSingle(
     collectionName,
     id: id,
     mapper: _mapper,
-    expansions:
-        (baseExpansions ?? {})..addAll(additionalExpansions ?? const {}),
+    fields: fields,
+    expansions: (expansions ?? {})..addAll(additionalExpansions ?? const {}),
     query: query,
     headers: headers,
   );
@@ -180,14 +190,15 @@ class CollectionHelper<T extends PocketBaseRecord> {
   Future<T?> getMaybeSingle(
     String id, {
     Map<String, String>? additionalExpansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) => _helper.getMaybeSingle(
     collectionName,
     id: id,
     mapper: _mapper,
-    expansions:
-        (baseExpansions ?? {})..addAll(additionalExpansions ?? const {}),
+    expansions: (expansions ?? {})..addAll(additionalExpansions ?? const {}),
+    fields: fields,
     query: query,
     headers: headers,
   );
@@ -196,13 +207,14 @@ class CollectionHelper<T extends PocketBaseRecord> {
   Future<T> create({
     required Map<String, dynamic> data,
     Map<String, String>? additionalExpansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) => _helper.create(
     collectionName,
     data: data,
-    expansions:
-        (baseExpansions ?? {})..addAll(additionalExpansions ?? const {}),
+    expansions: (expansions ?? {})..addAll(additionalExpansions ?? const {}),
+    fields: fields,
     mapper: _mapper,
     query: query,
     headers: headers,
@@ -212,14 +224,15 @@ class CollectionHelper<T extends PocketBaseRecord> {
   Future<T> update(
     T record, {
     Map<String, String>? additionalExpansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) => _helper.update(
     collectionName,
     record: record,
     mapper: _mapper,
-    expansions:
-        (baseExpansions ?? {})..addAll(additionalExpansions ?? const {}),
+    expansions: (expansions ?? {})..addAll(additionalExpansions ?? const {}),
+    fields: fields,
     query: query,
     headers: headers,
   );
@@ -235,8 +248,15 @@ class CollectionHelper<T extends PocketBaseRecord> {
   /// - record id
   /// - filename
   /// and returns a correct Uri that can be used to retrieve it form the database
-  Uri getFileUri(String id, String filename) {
-    return pb.buildURL('api/files/$collection/$id/$filename');
+  Uri getFileUrl(
+    String id,
+    String filename, {
+    Map<String, dynamic>? queryParameters,
+  }) {
+    return pb.buildURL(
+      'api/files/$collection/$id/$filename',
+      queryParameters ?? {},
+    );
   }
 
   ///Add files to a record, this takes:
@@ -245,13 +265,17 @@ class CollectionHelper<T extends PocketBaseRecord> {
   Future<T> addFiles(
     String id, {
     required Map<String, Uint8List> files,
+    Map<String, String>? additionalExpansions,
+    List<String>? fields,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) => _helper.addFiles(
     collectionName,
     id: id,
     files: files,
+    fields: fields,
     mapper: _mapper,
+    expansions: (expansions ?? {})..addAll(additionalExpansions ?? const {}),
     query: query,
     headers: headers,
   );
@@ -262,6 +286,7 @@ class CollectionHelper<T extends PocketBaseRecord> {
   Future<T> removeFiles(
     String id, {
     required List<String> fileNames,
+    Map<String, String>? additionalExpansions,
     Map<String, dynamic>? query,
     Map<String, String>? headers,
   }) => _helper.removeFiles(
@@ -269,6 +294,7 @@ class CollectionHelper<T extends PocketBaseRecord> {
     id: id,
     fileNames: fileNames,
     mapper: _mapper,
+    expansions: (expansions ?? {})..addAll(additionalExpansions ?? const {}),
     query: query,
     headers: headers,
   );